+++ packages/frontend/editor-ui/src/composables/useExpressionEditor.test.ts
@@ -9,7 +9,6 @@ import { fireEvent, waitFor } from '@testing-library/vue';
 import { setActivePinia } from 'pinia';
 import { beforeEach, describe, vi } from 'vitest';
 import { defineComponent, h, ref, toValue } from 'vue';
-import { useRouter } from 'vue-router';
 import { useExpressionEditor } from './useExpressionEditor';
 
 vi.mock('@/composables/useAutocompleteTelemetry', () => ({
