+++ packages/frontend/editor-ui/src/composables/useWorkflowHelpers.test.ts
@@ -737,58 +737,4 @@ describe('useWorkflowHelpers', () => {
 			expect(result.source).toBeNull();
 		});
 	});
-
-	describe('saveCurrentWorkflow', () => {
-		beforeEach(() => {
-			setActivePinia(createTestingPinia({ stubActions: false }));
-
-			workflowsStore = mockedStore(useWorkflowsStore);
-
-			nodeTypesStore = mockedStore(useNodeTypesStore);
-			nodeTypesStore.setNodeTypes(nodeTypes);
-		});
-
-		it('should save the current workflow', async () => {
-			const workflow = createTestWorkflow({
-				id: 'w0',
-				nodes: [createTestNode({ type: CHAT_TRIGGER_NODE_TYPE, disabled: false })],
-				active: true,
-			});
-
-			vi.spyOn(workflowsStore, 'fetchWorkflow').mockResolvedValue(workflow);
-			vi.spyOn(workflowsStore, 'updateWorkflow').mockResolvedValue(workflow);
-
-			workflowsStore.setWorkflow(workflow);
-
-			const { saveCurrentWorkflow } = useWorkflowHelpers();
-			await saveCurrentWorkflow({ id: 'w0' });
-			expect(workflowsStore.updateWorkflow).toHaveBeenCalledWith(
-				'w0',
-				expect.objectContaining({ id: 'w0', active: true }),
-				false,
-			);
-		});
-
-		it('should include active=false in the request if the workflow has no activatable trigger node', async () => {
-			const workflow = createTestWorkflow({
-				id: 'w1',
-				nodes: [createTestNode({ type: CHAT_TRIGGER_NODE_TYPE, disabled: true })],
-				active: true,
-			});
-
-			vi.spyOn(workflowsStore, 'fetchWorkflow').mockResolvedValue(workflow);
-			vi.spyOn(workflowsStore, 'updateWorkflow').mockResolvedValue(workflow);
-
-			workflowsStore.setWorkflow(workflow);
-
-			const { saveCurrentWorkflow } = useWorkflowHelpers();
-			await saveCurrentWorkflow({ id: 'w1' });
-			expect(workflowsStore.updateWorkflow).toHaveBeenCalledWith(
-				'w1',
-				expect.objectContaining({ id: 'w1', active: false }),
-				false,
-			);
-			expect(workflowsStore.setWorkflowInactive).toHaveBeenCalled();
-		});
-	});
 });
